fix create temporary directory

+++ src/agat/agat_sp_merge_annotations/test.sh
@@ -8,9 +8,9 @@ set -eo pipefail
 test_dir="${meta_resources_dir}/test_data"
 
 # create temporary directory and clean up on exit
-TMPDIR=$(mktemp --tmpdir "$meta_temp_dir")
+TMPDIR=$(mktemp -d "$meta_temp_dir/$meta_functionality_name-XXXXXX")
 function clean_up {
-  rm -rf "$TMPDIR"
+ [[ -d "$TMPDIR" ]] && rm -rf "$TMPDIR"
 }
 trap clean_up EXIT
